rename: useFetchUserInfo 훅 파일 위치 변경 (#37)
